backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.30
alembic>=1.13.0
psycopg2-binary>=2.9.9
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.1
python-json-logger>=2.0.7
httpx>=0.27.0
sslyze>=6.0.0
cryptography>=42.0.0
cyclonedx-python-lib>=7.0.0
numpy>=1.26.0
scipy>=1.13.0
networkx>=3.3
pyjwt>=2.8.0
geoip2>=4.8.0
jinja2>=3.1.4
weasyprint>=62.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
rich>=13.7.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
